Requiring fresh token in API

+++ section_11_jwt_extended/resources/user.py
@@ -1,4 +1,11 @@
-from flask_jwt_extended import create_access_token, create_refresh_token
+from flask_jwt_extended import (
+    create_access_token, 
+    create_refresh_token, 
+    fresh_jwt_required,
+    get_jwt_claims,
+    get_jwt_identity,
+    jwt_refresh_token_required,
+)
 from flask_restful import Resource, reqparse, request
 from models.user import UserModel as User
 
@@ -13,20 +20,30 @@ class UserResource(Resource):
             }, 404
 
         return {
+            'code': 200,
             'user': user._json()
         }, 200
 
     # Menghapus data akun pengguna
     def delete(self, id):
+        claims = get_jwt_claims()
+        if claims['is_admin'] is None:
+            return {
+                'code': 403,
+                'message': 'You shall not pass!'
+            }, 403
+
         user = User.find(id)
         if user is None:
             return {
+                'code': 403,
                 'message': 'User not found'
             }, 400
 
         user.delete()
 
         return {
+            'code': 200,
             'message': 'User deleted successfully'
         }, 200
 
@@ -71,6 +88,7 @@ class UsersResource(Resource):
         user.save()
 
         return {
+            'code': 201,
             'message': 'User created successfully'
         }, 201
 
@@ -113,4 +131,21 @@ class UserAuth(Resource):
             },
             'access_token': access_token,
             'refresh_token': refrech_token
+        }, 200
+
+
+class RefreshToken(Resource):
+    # Generate new access token
+    @jwt_refresh_token_required
+    def get(self):
+        user_id = get_jwt_identity()
+        user = User.find(user_id)
+        access_token = create_access_token(identity=user_id, fresh=False)
+        return {
+            'user': {
+                'id': user.id,
+                'name': user.name
+            },
+            'access_token': access_token,
+            'refresh_token': None
         }
